requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
django-filter==2.3.0
djangorestframework==3.11.1
googlemaps==4.4.2
idna==2.10
prettyconf==2.1.0
PrettyConfig==0.0.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
drf-yasg==1.17.1
django==3.0.5
django-bootstrap-form==3.4
markdown==3.2.1
django-widget-tweaks==1.4.8
python-decouple==3.3
gunicorn==20.0.4